add override of auto-detected locale encoding
The files for the site.t test are in UTF-8. As long as we didn't care
what encoding we were using, it would work. But now that we're
encoding/decoding, we need to be conscious that our encoding guess could
be wrong.

This prevents errors like `ascii "\xF0" does not map to Unicode` when
reading files.

Thanks @eserte for the report!

Fixes #10