fix compatibility with Yancy > v1.057
Yancy v1.057 removed an undocumented "feature" that was only being used
by the editor: The `id_field` stash value would need to be set to the
field that should be looked up from the `id` stash value. This wasn't
very elegant and prevented multiple ID fields, so it was changed to be
whatever stash value named for the ID field.

Thanks @tehmoth for reporting this issue! Fixes #11